the "about" panel was completed and refactor code
