+++ app/src/main/java/com/example/pokedex/ui/pokemonDetail/fragments/AboutFragment/view/AboutFragment.kt
@@ -1,6 +1,7 @@
-package com.example.pokedex.ui.pokemonDetail.fragments.AboutFragment
+package com.example.pokedex.ui.pokemonDetail.fragments.AboutFragment.view
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -8,9 +9,12 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.viewModels
 import com.example.pokedex.databinding.FragmentAboutBinding
+import com.example.pokedex.ui.pokemonDetail.fragments.AboutFragment.viewmodel.PokemonDetailViewModel
 import com.example.pokedex.utils.ViewState
 import com.graphqlapollo.PokemonDetailQuery
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class AboutFragment(name: String) : Fragment() {
     private var _binding: FragmentAboutBinding? = null
     private val binding get() = _binding!!
@@ -28,10 +32,8 @@ class AboutFragment(name: String) : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        Toast.makeText(activity, pokemonName, Toast.LENGTH_SHORT).show()
-
-        // observeData()
-        //viewModel.queryPokemonDetail(pokemonName)
+        observeData()
+        viewModel.queryPokemonDetail(pokemonName)
     }
 
     private fun observeData() {
@@ -41,6 +43,7 @@ class AboutFragment(name: String) : Fragment() {
                     val results = response.value?.data?.pokemon
                     if (results != null) {
                         pokemonDetail = results
+                        initPokemonData()
                     }
                 }
                 is ViewState.Error -> {
@@ -50,6 +53,26 @@ class AboutFragment(name: String) : Fragment() {
                 }
             }
         }
+    }
+
+    private fun initPokemonData() {
+        binding.tvPokemonHeight.text = pokemonDetail.height.toString()
+        binding.tvPokemonWeight.text = pokemonDetail.weight.toString()
+        var pokemonType = ""
+        pokemonDetail.types?.forEach { typeName ->
+            typeName!!.type?.name?.let { pokemonType += ", ${it.replaceFirstChar { it.uppercase()}} "}
+        }
+        binding.tvType.text = pokemonType.replaceFirstChar {""}
+        var pokemonAbilities = ""
+        pokemonDetail.abilities?.forEach { abilitiesName ->
+            pokemonAbilities += (", " + abilitiesName?.ability?.name) ?: ""
+        }
+        binding.tvAbilities.text = pokemonAbilities.replaceFirstChar { "" }
+        var pokemonMoves = ""
+        pokemonDetail.moves?.forEach{ moveName ->
+            pokemonMoves+= (", " + moveName?.move?.name) ?: ""
+        }
+        binding.tvMoves.text = pokemonMoves.replaceFirstChar { "" }
     }
 
 }
