+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,17 @@
 <resources>
-    <string name="app_name">Pokedex</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="app_name">Pokédex</string>
     <string name="about">About</string>
     <string name="stats">Stats</string>
     <string name="evolution">Evolution</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="pokedex_data">Pokémon Data</string>
+
+    <string name="pokemon_height">Height:</string>
+    <string name="pokemon_weight">Weight:</string>
+    <string name="pokemon_type">Type:</string>
+    <string name="pokemon_abilities">Abilities:</string>
+    <string name="pokemon_moves">Moves:</string>
+
+
+
 </resources>
