+++ build.gradle
@@ -15,8 +15,8 @@ buildscript {
 
 
 plugins {
-    id 'com.android.application' version '7.2.1' apply false
-    id 'com.android.library' version '7.2.1' apply false
+    id 'com.android.application' version '7.3.1' apply false
+    id 'com.android.library' version '7.3.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.10' apply false
     id 'org.jlleitschuh.gradle.ktlint' version '11.0.0'
     id 'com.google.dagger.hilt.android' version '2.41' apply false
